refactor: articleSection update actions
