+++ src/server/cms/articleSections/update.ts
@@ -8,6 +8,7 @@ import { createCmsLink } from '@/server/cms/links/create'
 import type { ImageSize, ArticleSection, Position } from '@prisma/client'
 import type { ExpandedArticleSection } from '@/cms/articleSections/Types'
 import type { ActionReturn } from '@/actions/Types'
+import type { ArticleSectionPart } from './Types'
 
 /**
  * This is the function that updates an article section metadata about how the (cms)image is displayed
@@ -55,6 +56,158 @@ export async function updateArticleSection(nameOrId: string, changes: {
             include: articleSectionsRealtionsIncluder
         })
         return { success: true, data: articleSection }
+    } catch (error) {
+        return createPrismaActionError(error)
+    }
+}
+
+/**
+ * This is the function that adds a part to an article section. The part can be a cmsImage, cmsParagraph.
+ * If the part already exists in the article section, it will return an error
+ * If it does not it will create the part and add it to the article section
+ * @param nameOrId - The name or id of the article section to add the part to
+ * @param part - The part to add to the article section cmsImage, cmsParagraph or cmsLink
+ * @returns - The updated article section
+ */
+export async function addArticleSectionPart(
+    nameOrId: string, 
+    part: ArticleSectionPart
+): Promise<ActionReturn<ExpandedArticleSection>> {
+    const where = {
+        name: typeof nameOrId === 'string' ? nameOrId : undefined, 
+        id: typeof nameOrId === 'number' ? nameOrId : undefined
+    } as const
+
+    try {
+        const articleSection = await prisma.articleSection.findUnique({
+            where,
+            include: { [part]: true }
+        })
+        if (!articleSection) {
+            return createActionError('NOT FOUND', 'ArticleSection not found')
+        }
+        if (articleSection[part]) {
+            return createActionError('BAD PARAMETERS', `ArticleSection already has ${part}`)
+        }
+
+        switch (part) {
+            case 'cmsImage':
+            {
+                const cmsImage = await createCmsImage(`${nameOrId}_image`)
+                if (!cmsImage.success) return cmsImage
+                return {
+                    success: true,
+                    data: await prisma.articleSection.update({
+                        where,
+                        data: { cmsImage: { connect: { id: cmsImage.data.id } } },
+                        include: articleSectionsRealtionsIncluder
+                    })
+                }
+            }
+            case 'cmsParagraph':
+            {
+                const cmsParagraph = await createCmsParagraph(`${name}_paragraph`)
+                if (!cmsParagraph.success) return cmsParagraph
+                return {
+                    success: true,
+                    data: await prisma.articleSection.update({
+                        where,
+                        data: { cmsParagraph: { connect: { id: cmsParagraph.data.id } } },
+                        include: articleSectionsRealtionsIncluder
+                    })
+                }
+            }
+            case 'cmsLink':
+            {
+                const cmsLink = await createCmsLink(`${name}_link`)
+                if (!cmsLink.success) return cmsLink
+                return {
+                    success: true,
+                    data: await prisma.articleSection.update({
+                        where,
+                        data: { cmsLink: { connect: { id: cmsLink.data.id } } },
+                        include: articleSectionsRealtionsIncluder
+                    })
+                }
+            }
+            default:
+                break
+        }
+        return createActionError('BAD PARAMETERS', 'Invalid part')
+    } catch (error) {
+        return createPrismaActionError(error)
+    }
+}
+
+/**
+ * This is a function that removes a part from an article section. The part can be a cmsImage, cmsParagraph.
+ * It does so by deleting the associated part from the database, in tern setting the part to null in the article section
+ * If the attribute destroyOnEmpty is true, it will also remove the article section if all parts are removed
+ * @param nameOrId - The name or id of the article section to remove the part from
+ * @param part - The part to remove from the article section cmsImage, cmsParagraph or cmsLink
+ * @returns - The updated article section
+ */
+export async function removeArticleSectionPart(
+    nameOrId: string, 
+    part: ArticleSectionPart
+): Promise<ActionReturn<ArticleSection>> {
+    const where = {
+        name: typeof nameOrId === 'string' ? nameOrId : undefined, 
+        id: typeof nameOrId === 'number' ? nameOrId : undefined
+    } as const
+    try {
+        const articleSection = await prisma.articleSection.findUnique({
+            where,
+            include: { cmsLink: true, cmsParagraph: true, cmsImage: true }
+        })
+        if (!articleSection) {
+            return createActionError('NOT FOUND', 'ArticleSection not found')
+        }
+        if (!articleSection[part]) {
+            return createActionError('BAD PARAMETERS', `ArticleSection does not have ${part}`)
+        }
+
+        switch (part) {
+            case 'cmsLink':
+                await prisma.cmsLink.delete({ where: { id: articleSection.cmsLink?.id } })
+                break
+            case 'cmsParagraph':
+                await prisma.cmsParagraph.delete({ where: { id: articleSection.cmsParagraph?.id } })
+                break
+            case 'cmsImage':
+                await prisma.cmsImage.delete({ where: { id: articleSection.cmsImage?.id } })
+                break
+            default:
+                break
+        }
+
+
+        // check if all Parts are removed and if so, remove the articleSection,
+        // but only if destroyOnEmpty is true
+        const afterDelete = await prisma.articleSection.findUnique({
+            where,
+            include: { cmsParagraph: true, cmsImage: true, cmsLink: true }
+        })
+        if (!afterDelete) {
+            return createActionError('UNKNOWN ERROR', 'Noe uventet skjedde etter sletting av del av artclesection')
+        }
+        if (
+            articleSection.destroyOnEmpty &&
+            !afterDelete.cmsImage &&
+            !afterDelete.cmsParagraph &&
+            !afterDelete.cmsImage
+        ) {
+            const destroyRes = await destroyArticleSection(name)
+            if (!destroyRes.success) {
+                return createActionError('UNKNOWN ERROR', 'Greide ikke slette artikkelseksjonen')
+            }
+            return { success: true, data: destroyRes.data }
+        }
+
+        return {
+            success: true,
+            data: afterDelete
+        }
     } catch (error) {
         return createPrismaActionError(error)
     }
